+++ packages/shared/index.css
@@ -1,1 +1,1 @@
-@import './src/variable.css';
+@import './src/designVariables/variable.css';
