+++ packages/shared/src/designVariables/index.ts
@@ -1,0 +1,4 @@
+import TYPOGRAPHY from './typography'
+import VARIABLES from './variables'
+
+export { TYPOGRAPHY, VARIABLES }
